requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
pandas==2.2.1
python-dotenv==1.0.1
fake-useragent==1.5.1
streamlit==1.32.0
dateparser==1.2.0
